pull updates

+++ install.ps1
@@ -1,46 +1,53 @@
-# iwr -useb https://secman-team.github.io/install/install.ps1 | iex
+# iwr -useb https://secman-team.github.io/install.ps1 | iex
 
-$loc = "$HOME\AppData\Local\secman"
-$smShUrl = "https://raw.githubusercontent.com/secman-team/tools/HEAD/sm.sh"
-$sm_winLoc = "$HOME\sm"
+if ((Get-Command ruby -errorAction SilentlyContinue) -or (Get-Command git -errorAction SilentlyContinue)) {
+  # get latest release
+  $release_url = "https://api.github.com/repos/secman-team/secman/releases"
+  $tag = (Invoke-WebRequest -Uri $release_url -UseBasicParsing | ConvertFrom-Json)[0].tag_name
 
-if (Test-Path -path $loc) {
-  Remove-Item $loc -Recurse -Force
-}
+  $loc = "$HOME\AppData\Local\secman"
+  $smPS1URL = "https://raw.githubusercontent.com/secman-team/tools/HEAD/sm.ps1"
+  $sm_winLoc = "$HOME\sm"
 
-if (Test-Path -path $sm_winLoc) {
-  Remove-Item $sm_winLoc -Recurse -Force
-}
+  if (Test-Path -path $loc) {
+    Remove-Item $loc -Recurse -Force
+  } elseif (Test-Path -path $sm_winLoc) {
+    Remove-Item $sm_winLoc -Recurse -Force
+  }
 
-Invoke-WebRequest $smShUrl -outfile sm.sh
+  Write-Host "Installing secman..." -ForegroundColor DarkCyan
 
-$lv = bash sm.sh
+  Invoke-WebRequest https://github.com/secman-team/secman/releases/download/$tag/secman_windows_${tag}_x64.zip -outfile secman_windows.zip
 
-Write-Host "Installing secman..." -ForegroundColor DarkCyan
+  Expand-Archive secman_windows.zip
 
-Invoke-WebRequest https://github.com/secman-team/secman/releases/download/$lv/secman_windows_${lv}_x64.zip -outfile secman_windows.zip
+  New-Item -ItemType "directory" -Path $loc
 
-Expand-Archive secman_windows.zip
+  Move-Item -Path secman_windows\bin -Destination $loc
 
-New-Item -ItemType "directory" -Path $loc
+  git clone https://github.com/secman-team/sm-win $sm_winLoc
 
-Move-Item -Path secman_windows\bin -Destination $loc
+  Invoke-WebRequest $smPS1URL -outfile $sm_winLoc\sm.ps1
 
-git clone https://github.com/secman-team/sm-win $sm_winLoc
+  Write-Host "Installing ruby deps..." -ForegroundColor DarkYellow
 
-Invoke-WebRequest $smShUrl -outfile $sm_winLoc\sm.sh
+  gem install colorize
 
-Write-Host "Installing ruby deps.." -ForegroundColor DarkYellow
+  Remove-Item secman_windows* -Recurse -Force
 
-gem install colorize optparse
+  [System.Environment]::SetEnvironmentVariable("Path", $Env:Path + ";$loc\bin", [System.EnvironmentVariableTarget]::User)
 
-Remove-Item secman_windows* -Recurse -Force
-Remove-Item sm.sh
+  Write-Host "Installing @secman/sm-upg" -ForegroundColor DarkBlue
 
-if (Test-Path -path $loc) {
-  Write-Host "Yessss, secman was installed successfully, run secman --help"  -ForegroundColor DarkGreen
+  npm i -g @secman/sm-upg
+
+  if (Test-Path -path $loc) {
+    Write-Host "Yessss, secman was installed successfully, run secman --help" -ForegroundColor DarkGreen
+  } else {
+    Write-Host "Download failed 😔"
+  }
 } else {
-  Write-Host "Download failed 😔"
+  Write-Host "Some of these apps must be installed: git, or ruby"
+  Write-Host "git: https://git-scm.com"
+  Write-Host "ruby: https://rubyinstaller.org"
 }
-
-[System.Environment]::SetEnvironmentVariable("Path", $Env:Path + ";$loc\bin", [System.EnvironmentVariableTarget]::User)
